requirements.txt
amqp==2.6.1
argh==0.26.2
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==4.4.7
certifi==2023.7.22
charset-normalizer==2.1.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==3.2.7
django-appconf==1.0.5
django-celery-email==3.0.0
django-environ==0.7.0
djangorestframework==3.14.0
flower==0.9.7
google-play-scraper==1.2.4
humanize==4.8.0
idna==3.4
kombu==4.6.11
lxml==4.5.2
prometheus-client==0.8.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==5.3.1
redis==3.5.3
requests==2.28.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tornado==6.3.3
typing-extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
vine==1.3.0
watchdog==1.0.2
wcwidth==0.2.6
